It's not very nice or neat, but it works =)
The idea for a store website has been implemented.

On the main page "Home" ten random products are displayed by clicking on which you can go and see full information.

Next, after “Home”, sections with categories, clicking on one of the categories redirects the User to a page with products from the selected category.

Products can be sorted alphabetically, by price and by rating. There is also an opportunity to enlarge the product card.

When clicking on one of the products, the User is transferred to a page with a detailed description of the product. Additionally, the ability to click on a product category to view products from this category has been added.

There is also a search option. After entering a keyword and clicking on the search button, the User is transferred to a page with search results.